removed old nav bar, moved search page into tabs directory, and fixed some design issues in the filter modal
